refactor(type): move context definitions into separate modules
